New driver ManagementInterface
This blueprint will introduce a new driver interface called Management
Interface, this interface will support management operations like:
Setting the boot device, getting the current boot device and getting
the supported boot devices. Of course this is just the initial methods,
and more operations could be added in the future.

Change-Id: I96c7e8f75c67a71544958aec60c9d13fb4212b0b