API to get driver_info properties
It is possible to create a node without specifying any driver_info properties
in the initial POST -- this is reasonable and fine. However, the API does not
expose what driver_info properties are required for the node's driver. The
client cannot know what fields/properties to send in subsequent PATCH requests
without reading Ironic's developer docs (or source code!).

To address the above, an API is proposed, that returns the driver_info
properties, along with a description for each property. The description will
include whether the property is required or not.

This is a specification for blueprint
https://blueprints.launchpad.net/ironic/+spec/get-required-driver-info.

Change-Id: I5e1dd5d0ee8c71bf2e81d4ce074296c6c4e15220