Set up for kilo specs
This adds the entries in the index so that the main page will
list the specs in specs/kilo. To get sphinx to not err, a dummy
.rst file (based on the template.rst) was added to specs/kilo.
It should be deleted after we have an approved spec.

Without this temporary fix, all spec patches will basically
need to have the change in index.rst; otherwise Jenkins will
not be happy.

Note that the dummy .rst file needs to be added or sphinx will
complain because the specs/kilo/* pattern didn't match any
documents. Furthermore, the .rst file there must be valid
(based on the template) or the unit tests will fail. And lastly,
that dummy .rst file cannot be named 'template.rst' because
we've configured it to ignore all files named 'template.rst'. If
that is ignored, we're back to no files in specs/kilo, and sphinx
will be unhappy again.

Closes-Bug: #1379415
Change-Id: Ic4a40a5b085073bc9eebc3d2c75432f18f01cbef