Adjust the extended-vendor-passthru spec
The way the merged spec described how the vendor methods will be
exposed via our API doesn't work because the 'method' to be invoked
in the vendor_passthru is not an endpoint it's a parameter. For e.g:

GET {'method': <method name>} ../vendor_passthru
or
GET ../vendor_passthru/<method name>

Is the same thing, so once we extend the vendor_passtru endpoint to
support other HTTP methods (including GET) we can't use it also
to expose the supported methods.

That said I'm changing the proposal here to add a new endpoint to
the vendor_passthru that can be used for exposing the supported
methods:

GET ../vendor_passthru/methods

APIImpact
Change-Id: I0dc02848e8c722945127877b859b935fa482e85f