Check that filename is same as blueprint name
The template was updated to indicate that the filename of the
specification must match the name of the blueprint (changed the
name of the sample blueprint be 'template' to match template.rst),
and that the title of the specification could be the title of the
blueprint.

A unit test was added to check the filename against the blueprint name.

Fixed a typo in one of the unit tests.

Change-Id: Ia5180aa44d6b2967d4843b8ab176915181e5e77f