Correct the diagram for the "current" state machine
After working on a formal model of the current state machine, it became
apparent that the diagram in this spec did not match the behavior in the
Juno release. The representation of "momentary" states is misleading --
these were never saved in or written to the node.provision_state field,
and were instead just return values from function calls.

As the diagram was thus misleading, this patch replaces the diagram
based on the actual behavior in the Juno release.

Change-Id: I5882e2ef4d0d63e8dcc01e9c06d7343bf35473a3