Rename MANAGED to MANAGEABLE
The spec says that states ending with -ING and -ED are active states
(momentary, states that don't need a API call to transition from). The
only exception is the MANAGED state which is a passive state (a state that
needs an API call to transition from). This patch is renaming MANAGED
to MANAGEABLE to get rid of that exception and make naming more consistent
in the new state machine.

Change-Id: I7224c18d6a1afa063249232496f6e34729d7b794