Change clean_steps to a dict
I hit a lot of ugly code trying to marshal the namedtuples into
the dictionary in the database and back again, so I'm just going to
use a dictionary the entire time.

While implementing cleaning, a debate came up about the best way to
store the current clean step and how to use that to coordinate
the next clean step. I decided that storing the whole dict of
the current clean step is the easiest way, and presents a better
API to the operator.

There was some confusion as to driver vs interface. Cleaning is
implemented by the interfaces, rather than the driver as a whole,
so I fixed the terminology in the spec.

Change-Id: I99b99e7760f44683447fce1f227baaabc494874d