Open specs for Liberty
This updates the template to ask about microversions, as well as
creating the liberty/ dir with a placeholder spec which should be
removed once we have merged an approved spec for Liberty.

Additionally, all specs from Kilo that were not implemented have
been archived. If they wish to be revived for Liberty, a separate
merge request should be issued moving them into liberty/.

Change-Id: I839b478a67af9eab3bfefbf979227e5f49b01aed