Add a 'Client (CLI) impact' section to the spec
This adds a 'Client (CLI) impact' section to the specification template.
The place to capture this information was in 'Other end user impact' but
that isn't explicit enough for people to know to put that information
there or for the reviewers to look there for that information.

Having an explicit section will also make it easier later, to check
that the feature proposed in the specification has been completely
implemented.

Change-Id: If8bec359034d35b4edf32f473645c1c87a9fe479