Update RAID spec to remove some configurable defaults
This commit removes the CONF variables
CONF.raid.share_physical_disks, CONF.raid.disk_type and
CONF.raid.interface_type.  This is because in heterogenous
hardware environments, it is not possible to provide a single
disk_type and interface_type as the default value as that will
lead to cleaning failures (when it is unable to find suitable
devices).  If operator wants to use specific type of disks for a
bare metal node, they should rather set it in node's target
raid config. It also removes CONF.raid.share_physical_disks as it
is not common to share physical disks across RAID disks (and if
required can again be set to False in the node's target_raid_config).

Change-Id: Ia5e2c2774c71d68506742edd15f68a37f1131761