Add placeholders for all previous specs
This commit adds a placeholder with link to the new location for each of
the specs that were moved by the previous commit. It also adds a
"placeholder" file containing TOC-style references to the old
directories, and a link to this on the index page.

This is broken into two commits to preserve the history; git can't
recognize a rename and replace operation in the same commit.

Change-Id: Ieea08e7b477993baaca76b14210547e3e26c2935
Co-authored-by: Jim Rollenhagen <jim@jimrollenhagen.com>