supported-boot-device-list placeholder (2 of 2)
Due to the reorg of the specs repo, the supported-boot-device
spec needs to live in the approved/, not the liberty/,
subdirectory. (With a placeholder in the old directory.)
This move will be done in two commits to preserve the history;
git can't recognize a rename and replace operation in the same commit.

Change-Id: Icb2bd97f91bef7148e0294b6582cef92f83c5739