Organize Liberty specs into implemented versions
Now that we have semver-versioned releases of ironic, it would be
useful to show which specifications were implemented in each
release.

New subdirectories 4.0 & 4.2 are created for each ironic release
during the Liberty development cycle (where any blueprints were
implemented) and symbolic links are added to the actual
specification files in liberty-implemented/.

The specifications at liberty-implemented/ are still generated and
available at that older URL location, but we don't put any explicit
links to those older URLs in the index.

Change-Id: I3e74d6055c4d56a56e54d0f34b53e0a695c3bb50