Active Node Creation
The workflow of ironic presently forces all new nodes to enter a
default initial state. From an operator perspective that is looking
to implement ironic to manage an evolving fleet of machines, it is
ideal to be able to create new nodes that are already in use, that
ironic can know about, to represent pre-existing systems, until
the operator releases the node back into the available pool.

Partial-Bug: #1526315
Change-Id: Id30e6fd35cac07aa78b1f72aebc9decd2b2e678b