Boot from Volume - Reference Drivers
During the Vancouver Summit, and most recently the ironic Mitaka
midcycle, the question of booting from a remote volume arose.

The decision from the ironic Mitaka midcycle was for a specification
to be drafted for reference driver changes. As such, this change
proposes that specification, in accordance with the prophecy.

Co-Authored-By: Chris Krelle <nobodycam@gmail.com>
Co-Authored-By: stephane <stephane@alum.mit.edu>

Partial-Bug: #1559691
Change-Id: I2d7607d967b4191b94b261005519ae46637e790d