Update active node creation spec based on feedback
Initially the active node creation specification, upon discussion
in the ironic community, it was desired to conceal the new states
from older clients. The desire to do this has changed.

Due to the review feedback from the revision to implement this
feature, https://review.openstack.org/#/c/275766/, I am revising
the specification.

Change-Id: Ie114bfaab249d73ea3ca7c0edc314ca1ed0448eb