Correct RAID approved spec JSON
The original example dictionaries added to the RAID spec
used single quotes instead of double quotes. Since double quotes
are used in JSON, as opposed to dictionaries, to represent a string,
they were failing JSON validation and could not be directly used.

Changed the single quotes to double quotes where appropriate
in the RAID spec and passed each example through a JSON
validator which now returns that the example is valid JSON.

Change-Id: I11c7258d30194d7841d5b0312d45243dee9953d6
Partial-bug: #1586157