API Evolution - ETAG identifiers
This specification describes the implementation of ETAG identifiers.

This patch is part of a series of patches which propose several
discrete units of work, broken out into separate specifications for
tracking. These have grown out of the discussions around whether
we should have a 'v2' API, and what current pain-points we all
experience with the current API. It is my belief, after
investigating all of the collected concerns, that we can implement
many of the desired changes incrementally and without a
full 'v2' rewrite.

Change-Id: Id8dc039fbf56f7e74dd79256b7a223f6eb673ad2
Related-bug: #1605728