Remove references to ZAP* state
This patch is removing all the references to the ZAP* states in the
spec. These states never got to be implemented and they were replaced
with CLEAN* states.

There's a link to this spec in the documentation about the states in
Ironic [0] and, leaving the reference for ZAP* states can confuse the
readers.

[0] http://docs.openstack.org/developer/ironic/dev/states.html

Change-Id: I859367f079e9e091c003a828500a49cbbbd11955