Run spec validation in linters job instead of py27
The default py27 job ignores the changes to *.rst files. The
linters jobs does not, and it makes more sense to include the
validation we do to the specs in it.

py27 is still left in the list of tox envs so that zuul does not
complain about it missing. It will be removed when we switch py27
to linters in project-config.

Change-Id: Ib14c2a3318dc171be8c8ed8207524ab9774c5ccd