Introduce inspect wait state
The in-band inspection is an asynchronous process isn't currently
handled through a "waiting" state. This is a discrepancy from the
rest of the asynchronous ironic states and may comprise a problem
for future features such as aborting the introspection or merging
ironic and ironic-inspector.

The spec adds a new inspect wait state to ironic state machine to
keep consistent with other asynchronous states.

Change-Id: I6fe864079d85fe7e7ba9650a51628e88e8a662cc
Partial-Bug: #1725211