Expose supported power states
This SPEC proposes adding a new API to expose supported power states of
nodes. This API would be helpful to see if a power action is supported
by a node because it depends on whether a power interface supports
power actions such as soft power off and soft reboot.

Change-Id: If4a3994376cd26a38ef4717bc215cd74df3591a8
Partial-Bug: 1734827