Use node.fault field for power fault recovery work
This updates the spec to use a node.fault field as opposed to
a node.maintenance_type field, to keep track of any fault
that is detected by ironic. (ironic will still put the node
in maintenance mode upon detection of a fault.)

A few other changes were made to the spec to clarify things.

Story: #1596107
Change-Id: I675c46e068654019073f5cb3bf4c290291ba405f