Moving UCS specs to backlog
Checked-in with Sam Betts and he indicated that these
items are not current priorites. As such, since they are
not being worked on, moving to backlog.

They can be moved back at any time, this move is to try
and help clarify what is in-progress or partially implemented
when looking at not-implemented.

Change-Id: Ice9e05806dea9c6409b8f52d20d7826a7ffb69bc