Amendment of the agent http provisioning spec
During spec[0] implementation, there are several findings related to
performance. This patch updates the image conversion rule to avoid
the situation when stream_raw_images and force_raw_images are both set
to true. Some descriptive text are added with regards to the options
that may affect performance.

[0] http://specs.openstack.org/openstack/ironic-specs/specs/not-implemented/agent-http-provisioning.html

Change-Id: I26179155f613e0bb694aa202094e0c52c0528131
Story: #1598852
Task: #10474