Fix ironic-specs repo doc builds
Since the documentation PTI job config was implemented
in zuul, the local docs builds no longer matched the builds
performed by zuul in terms of treating warnings as errors.

Adds flag to treat warnings as errors, and addresses the
numerous formatting issues that resulted in errors.

In some cases, the footnote no longer made sense to retain
and was as such removed.

Change-Id: Ie80d9c4480507c52fd0f3eee41fd6478e4c5c758