Support node history
Currently ironic uses one last_error field to record error information
when an operation failed. The field is easily overwritten and we don't
have a traceback on what happened in the past, the only way is to check
service logs.

The proposal is to introduce a new table named node_history, and record
important node events that helps bare metal maintenance and troubleshooting.

Change-Id: I3b8832a945183ce3ed41ea79838fc9f682bfc547
Story: 2002980
Task: 22989