Amend the in-band deploy steps spec
The current spec leaves a too narrow priority space for custom
in-band steps, while leaving 70 numbers after the final instance
is booted, which is arguably not help. Spread the proposed steps
evenly between 100 and 20, leaving 3 blocks with 19 priorities
each for customization.

Also rename "remove_provisioning_network" since removing the network
is not the only thing it does.

Change-Id: I9861f4a8878789e2ae68e8e69c4e5660e33f4989