Enumerate retired specs
We've been keeping a backlog of specs which are entirely outdated
and will never be implemented. And we should also indicate when
a specification document is no longer applicable at all. As such
this change adds a retired folder, enumeration, and moves four
documents to the retired folder to indcate they are no longer
applicable to Ironic so end users have appropriate context should
they discover the documents in the future.

Change-Id: I1ab7d6726dc45283ca72df1df1d3c18e8ae8ea94