Framework for DPU management/orchustration
This specification revisits the original smartnic work
in order to attempt to move the needle to account for newer
technology which is on the horizon.

It is also important to note that supporting a model such
as is proposed in this spec, also addresses some of the
various discussions which have occured in the past few
years in terms of addressing more compelx models and
interactions by nesting the model and enablling the
nested model to be interacted with.

Change-Id: I57a2130da64056655fd57522ca76b8a2a727da88