Update standalone networking initial conditions
This updates the standalone networking spec to addresses issues related
to the initial starting conditions.  Specifically, there is a 'chicken-
and-egg' scenario that exists in that LLDP information is required
in order to know which switch ports need to be configured for a node
but if those switch ports are disabled by default there is no way to
know which switch ports to configure.

This updates the document to state that switch ports must be manually
configured ahead of time and must be setup on the inspection VLAN to
facilitate collecting LLDP information.

Related-Bug: 2113769
Change-Id: I9a5854ce81b0a1478af34604b4fe093ecdac6e01
Signed-off-by: Allain Legacy <alegacy@redhat.com>